requirements.txt
Django==3.2.6
django-seed==0.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
ipdb==0.13.9
mysqlclient==2.0.3
Pillow==8.3.1
pre-commit==2.14.0
pytest-cov==2.12.1
pytest-django==4.4.0
pytest==6.2.4
python-memcached==1.59
pytz==2021.1
requests==2.26.0
